+++ src/rbindlist.c
@@ -289,7 +289,8 @@ SEXP rbindlist(SEXP l, SEXP usenamesArg, SEXP fillArg, SEXP idcolArg)
       }
       SEXP thisCol = VECTOR_ELT(li, w);
       int thisType = TYPEOF(thisCol);
-      if (TYPEORDER(thisType)>TYPEORDER(maxType)) maxType=thisType;
+      // Use >= for #546 -- TYPEORDER=0 for both LGLSXP and EXPRSXP (but also NULL)
+      if (TYPEORDER(thisType)>=TYPEORDER(maxType) && !isNull(thisCol)) maxType=thisType;
       if (isFactor(thisCol)) {
         if (isNull(getAttrib(thisCol,R_LevelsSymbol))) error("Column %d of item %d has type 'factor' but has no levels; i.e. malformed.", w+1, i+1);
         factor = true;
@@ -480,9 +481,9 @@ SEXP rbindlist(SEXP l, SEXP usenamesArg, SEXP fillArg, SEXP idcolArg)
           writeNA(target, ansloc, thisnrow);  // writeNA is integer64 aware and writes INT64_MIN
         } else {
           bool coerced = false;
-          if (TYPEOF(target)==VECSXP && TYPEOF(thisCol)!=VECSXP) {
+          if ((TYPEOF(target)==VECSXP || TYPEOF(target)==EXPRSXP) && TYPEOF(thisCol)!=TYPEOF(target)) {
             // do an as.list() on the atomic column; #3528
-            thisCol = PROTECT(coerceVector(thisCol, VECSXP));
+            thisCol = PROTECT(coerceVector(thisCol, TYPEOF(target)));
             coerced = true;
           } // else coerces if needed within memrecycle; possibly with a no-alloc direct coerce
           const char *ret = memrecycle(target, R_NilValue, ansloc, thisnrow, thisCol);